* sanity.sh (basicb): Also remove CVSROOT/Emptydir at end of 	test.  Otherwise it affects the toplevel-9 test for remote.
CVS-ID: src/ChangeLog 1.1114
CVS-ID: src/sanity.sh 1.233

+++ src/sanity.sh
@@ -914,10 +914,16 @@ done"
 	  cd ..
 	  rmdir 1
 
-	  rm -rf test2
+	  if test "$keep" = yes; then
+	    echo Keeping ${TESTDIR} and exiting due to --keep
+	    exit 0
+	  fi
+
+	  rm -r test2
 
 	  rm -rf ${CVSROOT_DIRNAME}/first-dir
 	  rm -rf ${CVSROOT_DIRNAME}/second-dir
+	  rm -rf ${CVSROOT_DIRNAME}/CVSROOT/Emptydir
 	  rm -f ${CVSROOT_DIRNAME}/topfile,v
 	  ;;
